requirements.txt
langchain==1.1.0
langchain-openai==1.1.0
python-dotenv
pandas
requests